+++ src/tests/test_cases/unicode_literals.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+"""This is a module."""
+
+from __future__ import unicode_literals
+from .expected import Expectation
+
+
+expectation = Expectation()
+expect = expectation.expect
+
+
+def with_unicode_docstring_without_u():
+    r"""Check unicode: \u2611."""